+++ NEWS.md
@@ -6,6 +6,8 @@ the dosing including dose amount and route.
 
 # PKNCA 0.10.1
 
+* A new parameter `aucabove.trough.all` was added to calculate the NCA above the
+  trough concentration.
 * Testing updates were made to work with dplyr version 1.1.0 (fix #198)
 
 # PKNCA 0.10.0